Implement a rudimentary import/export for storage boxes
Clicking 'download' will create a .json file with the entire contents of the first box.

Clicking 'upload' will open a file picker to select a file to load into storage. If successful it will be added as an entirely new box which means it can't currently be seen!

There is virtually no validation on the import which means it's probably very easy to break the local storage at the moment.